backend/requirements.txt
flask
flask-cors
langgraph
langchain
langchain-openai
python-dotenv
langgraph-checkpoint-sqlite
